+ AdviserHandled + SkippedProperty

+++ src/main/java/com/example/produitapi/models/dto/CommandeLigneDTO.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class CommandeLigneDTO {
 
     private ProduitDTO produit;
-    private int qtt;
+    private long qtt;
 
 }
